fix(ui): preserve initial prompt projection contracts
