+++ packages/gateway-client/src/browser.ts
@@ -7,7 +7,6 @@ export * from "./connect-auth.js";
 export * from "./protocol-client.js";
 export * from "./reconnect-policy.js";
 export * from "./session-projection.js";
-export * from "./session-projection-run-event.js";
 export * from "./session-subscriptions.js";
 export {
   DEFAULT_GATEWAY_REQUEST_TIMEOUT_MS,
