+++ packages/gateway-client/src/session-projection.ts
@@ -1,6 +1,6 @@
 /** Browser-safe identity and replay rules shared by Gateway conversation clients. */
 
-export { reduceSessionProjectionRunEvent } from "./session-projection-run-event.js";
+import { reduceSessionProjectionRunEventImpl } from "./session-projection-run-event.js";
 
 export type SessionMessageEnvelope = {
   messageId?: unknown;
@@ -48,6 +48,17 @@ export type SessionProjectionRun = {
   stopReason?: string;
   errorKind?: string;
   errorMessage?: string;
+};
+
+export type SessionProjectionGatewayRunEvent = {
+  state?: unknown;
+  yielded?: unknown;
+} & Partial<Record<"runId" | "message" | "stopReason" | "errorKind" | "errorMessage", unknown>>;
+
+export type SessionProjectionRunTransition = {
+  projection: SessionProjectionState;
+  previousRun: SessionProjectionRun | undefined;
+  currentRun: SessionProjectionRun | undefined;
 };
 
 export type SessionProjectionEntry = {
@@ -721,4 +732,13 @@ export function reduceSessionProjection(
     default:
       return state;
   }
+}
+
+/** Normalizes Gateway run envelopes once for every browser and terminal adapter. */
+export function reduceSessionProjectionRunEvent(
+  projection: SessionProjectionState,
+  event: SessionProjectionGatewayRunEvent,
+  scope: SessionProjectionScope = {},
+): SessionProjectionRunTransition | null {
+  return reduceSessionProjectionRunEventImpl(projection, event, scope);
 }
